Extract types to shared
