+++ backend/src/types/index.ts
@@ -1,16 +1,2 @@
-export * from './users';
-export * from './characters';
-export * from './stats';
-export * from './goals';
-export * from './family';
-export * from './tags';
-export * from './focus';
-export * from './todos';
-export * from './experiments';
-export * from './journals';
-export * from './quests';
-export * from './plans';
-export * from './xp';
-// Future exports:
-// export * from './posts';
-// export * from './auth';
+// Re-export all types from shared folder
+export * from '../../../shared/types';
