+++ frontend/src/lib/types/characters.ts
@@ -1,2 +1,2 @@
-// Import character types from backend (single source of truth)
-export type { Character, CreateCharacter, UpdateCharacter } from '../../../../backend/src/types/characters';
+// Import character types from shared folder (single source of truth)
+export type { Character, CreateCharacter, UpdateCharacter } from '../../../../shared/types/characters';
