+++ frontend/src/lib/types/focus.ts
@@ -1,2 +1,2 @@
-// Import focus types from backend (single source of truth)
-export type { Focus, NewFocus, FocusUpdate, CreateFocusRequest, UpdateFocusRequest, BatchUpdateFocusesRequest } from '../../../../backend/src/types/focus';
+// Import focus types from shared folder (single source of truth)
+export type { Focus, NewFocus, FocusUpdate, CreateFocusRequest, UpdateFocusRequest, BatchUpdateFocusesRequest } from '../../../../shared/types/focus';
