+++ frontend/src/lib/types/goals.ts
@@ -1,4 +1,4 @@
-// Import goal types from backend (single source of truth)
+// Import goal types from shared folder (single source of truth)
 export type {
   Goal,
   CreateGoal,
@@ -7,4 +7,4 @@ export type {
   CreateGoalWithTags,
   UpdateGoalWithTags,
   GoalWithParsedTags,
-} from '../../../../backend/src/types/goals';
+} from '../../../../shared/types/goals';
