+++ frontend/src/lib/types/tags.ts
@@ -1,2 +1,2 @@
-// Barrel file for tag types, re-exporting from backend
-export type { Tag, CreateTag, UpdateTag, TagWithCount, CreateTagWithName } from '../../../../backend/src/types/tags';
+// Barrel file for tag types, re-exporting from shared folder
+export type { Tag, CreateTag, UpdateTag, TagWithCount, CreateTagWithName } from '../../../../shared/types/tags';
