+++ frontend/src/lib/types/users.ts
@@ -1,2 +1,2 @@
-// Barrel file for user types, re-exporting from backend
-export type { User, NewUser, UserUpdate, PublicUser, UpdateUserRequest, UpdateUserAvatarRequest } from '../../../../backend/src/types/users';
+// Barrel file for user types, re-exporting from shared folder
+export type { User, NewUser, UserUpdate, PublicUser, UpdateUserRequest, UpdateUserAvatarRequest } from '../../../../shared/types/users';
