+++ frontend/src/lib/types/xpGrants.ts
@@ -1,4 +1,4 @@
-// Barrel file for XP grant types, re-exporting from backend
-import type { XpGrant, NewXpGrant, XpEntityType, XpSourceType, CreateXpGrantRequest, XpGrantFilter, XpGrantWithEntity } from '../../../../backend/src/types/xp';
+// Barrel file for XP grant types, re-exporting from shared folder
+import type { XpGrant, NewXpGrant, XpEntityType, XpSourceType, CreateXpGrantRequest, XpGrantFilter, XpGrantWithEntity } from '../../../../shared/types/xp';
 
 export type { XpGrant, NewXpGrant, XpEntityType, XpSourceType, CreateXpGrantRequest, XpGrantFilter, XpGrantWithEntity };
