+++ shared/types/index.ts
@@ -1,0 +1,15 @@
+// Shared types index - exports all types for both backend and frontend
+export * from './users';
+export * from './characters';
+export * from './stats';
+export * from './goals';
+export * from './family';
+export * from './tags';
+export * from './focus';
+export * from './todos';
+export * from './experiments';
+export * from './journals';
+export * from './journal-summaries';
+export * from './quests';
+export * from './plans';
+export * from './xp';
